+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [0.2.3] **NOT RELEASED**
 * Move `getLines` from `process.js` to `strings.js`
+* Add static method `getSignalName` to `Process` (`process.js`)
+* Add property `wasCancelled` to know whether or not a curl request was cancelled (`curl.js`)
+* Update documentation for `version` module
 
 ## [0.2.2]
 * Provide extra information in case of mismatch when using `tester.eq` (`tester.js)